requirements.txt
absl-py==2.3.1
accelerate==1.9.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
antlr4-python3-runtime==4.9.3
asttokens==3.0.1
async-timeout==5.0.1
attrs==25.3.0
certifi==2025.7.14
charset-normalizer==3.4.2
coloredlogs==15.0.1
comm==0.2.3
contourpy==1.3.2
cycler==0.12.1
debugpy==1.8.17
decorator==5.2.1
diffusers==0.34.0
easydict==1.13
einops==0.8.1
ema-pytorch==0.7.7
exceptiongroup==1.3.0
executing==2.2.1
filelock==3.18.0
flatbuffers==25.2.10
fonttools==4.59.0
frozenlist==1.7.0
fsspec==2025.7.0
ftfy==6.3.1
grpcio==1.74.0
hf-xet==1.1.5
huggingface-hub==0.34.0
imageio==2.37.0
imWatermark==0.0.2
matplotlib==3.10.3
numpy==2.2.6
omegaconf==2.3.0
onnxruntime==1.22.1
open_clip_torch==3.0.0
opencv-python==4.12.0.88
pandas==2.3.1
peft==0.18.0
pillow==11.3.0
pytorch-lightning==2.5.2
PyYAML==6.0.2
safetensors==0.5.3
timm==1.0.19
tokenizers==0.21.2
torch==2.7.1
torchmetrics==1.8.0
torchvision==0.22.1
tqdm==4.67.1
transformers==4.53.3
